Use configurable Gradle Inspector URL consistently
While it was possible to configure the URL of the repository to
download the Gradle Inspector through
`detect.gradle.inspector.repository.url`, checking for the most recent
version was still done through Maven Central. As Maven Central might
not be available in some setups (e.g. due to corporate policy),
hub-detect should not rely access Maven Central in case a custom URL
for the repository is configured.

In order to simplify the code and the init.gradle template and to make
this a non-breaking change, the Maven Central URL is now the default of
the `detect.gradle.inspector.repository.url` property, while this
property is also used to do the version resolution.

This might be a (partial) fix for #349.